pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "odrepomon"
version = "0.1.0"
description = "Gitignore-aware one-way mirror tool for backing up repos to OneDrive"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pathspec>=0.12.1",
  "PyYAML>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
mirror = "odrepomon.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
